+++ server/services/user.js
@@ -443,7 +443,9 @@ class UserService {
         .toFile(`${dir}/${imgFileName}`, function (err) {
           if (err) console.log(err);
         });
-      fs.unlinkSync(`${dir}/${oldImgFileName}`);
+      if (fs.existsSync(`${dir}/${oldImgFileName}`)) {
+        fs.unlinkSync(`${dir}/${oldImgFileName}`);
+      }
       return true;
     } else {
       return false;
